requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
geographiclib==1.52
    # via geopy
geopy==2.2.0
    # via -r requirements.in
gpxpy==1.5.0
    # via -r requirements.in
imageio==2.25.0
    # via -r requirements.in
lxml==4.9.1
    # via -r requirements.in
numpy==1.24.1
    # via
    #   -r requirements.in
    #   imageio
    #   opencv-python
    #   pandas
opencv-python==4.6.0.66
    # via -r requirements.in
pandas==1.5.3
    # via -r requirements.in
pillow==9.2.0
    # via imageio
python-dateutil==2.8.2
    # via pandas
pytz==2022.4
    # via pandas
six==1.15.0
    # via python-dateutil
tqdm==4.64.1
    # via -r requirements.in
